chore(i18n): convert listView hooks to TS
