+++ packages/plugins/i18n/admin/src/contentManagerHooks/tests/listView.test.ts
@@ -1,13 +1,15 @@
-import addColumnToTable from '../addColumnToTable';
+import { addColumnToTableHook } from '../listView';
 
-describe('i18n | contentManagerHooks | addColumnToTable', () => {
+describe('addColumnToTableHook', () => {
   it('does nothing when there s no i18n.localized key in the action', () => {
     const displayedHeaders = ['one'];
     const layout = {
+      components: {},
       contentType: { pluginOptions: {} },
     };
 
-    const result = addColumnToTable({ displayedHeaders, layout });
+    // @ts-expect-error – test purpose
+    const result = addColumnToTableHook({ displayedHeaders, layout });
 
     expect(result).toHaveProperty('displayedHeaders');
     expect(result).toHaveProperty('layout');
@@ -16,8 +18,9 @@ describe('i18n | contentManagerHooks | addColumnToTable', () => {
   });
 
   it('adds a header to the displayedHeaders array when the content type is localized', () => {
-    const displayedHeaders = [];
+    const displayedHeaders: unknown[] = [];
     const layout = {
+      components: {},
       contentType: {
         pluginOptions: {
           i18n: { localized: true },
@@ -25,7 +28,8 @@ describe('i18n | contentManagerHooks | addColumnToTable', () => {
       },
     };
 
-    const result = addColumnToTable({ displayedHeaders, layout });
+    // @ts-expect-error – test purpose
+    const result = addColumnToTableHook({ displayedHeaders, layout });
 
     // The anonymous function of cellFormatter creates problem, because it's anonymous
     // In our scenario, it's even more tricky because we use a closure in order to pass
